Add Basic Layout
This commit adds the basics for a Bash project. I have added
a bash library that I use a lot along with a docs directory.
This docs directory is also a common structure that I use.
This may be over kill for most Bash script projects. However,
I wanted to make it available.