Add Readme
I added a readme to the projects root because I feel
all good projects should have a readme.

I fixed the source path in lib/logging.sh. It was
pointing to a dir that does not exist in this
project.

I make change the mode of the script to be executable
because that is the whole point of the script.